+++ src/components/OurServices/OurServices.module.scss
@@ -8,10 +8,11 @@
   flex-direction: row;
   align-items: flex-end;
   @media (max-width: 768px) {
-    margin-top: 40px;
-    margin-bottom: 0px;
+    margin-top: 10px;
+    margin-bottom: -10px;
+    padding-top: 20px;
+    padding-bottom: 0px;
     width: 100%;
-    padding: 15px;
   }
   @media (min-width: 768px) and (max-width: 1024px) {
     margin-top: 40px;
@@ -73,7 +74,7 @@
   cursor: pointer;
   transition: all 0.3s ease-in-out;
   @media (max-width: 768px) {
-    width: 90%;
+    width: 100%;
   }
   @media (min-width: 768px) and (max-width: 1024px) {
     width: 32%;
