+++ src/components/SuportComponent/Support.module.scss
@@ -1,14 +1,28 @@
 .support {
-  margin: 100px auto 50px;
-  padding-top: 20px;
+  margin: 50px auto 50px;
+  padding-top: 50px;
   width: 78.9%;
   display: flex;
   flex-direction: row;
   justify-content: space-between;
+  @media (max-width: 768px) {
+    margin: 50px auto 50px;
+    padding-top: 40px;
+    padding-bottom: 40px;
+    width: 90%;
+    flex-direction: column;
+    justify-content: space-between;
+    align-items: center;
+    gap: 40px;
+  }
+  @media (min-width: 768px) and (max-width: 1024px) {
+    margin-top: 50px;
+    padding-top: 50px;
+    width: 100%;
+  }
 }
 
 .description {
-  margin-top: -20px;
   width: 27.3%;
   font-size: 17px;
   font-weight: 300;
@@ -28,25 +42,44 @@
   }
   @media (max-width: 768px) {
     text-transform: lowercase !important;
+    width: 90%;
   }
   @media (min-width: 768px) and (max-width: 1024px) {
     font-size: 16px !important;
+    width: 32%;
   }
 }
 
-.description::first-letter {
-  text-transform: capitalize;
-}
-
 .supportCard {
-  padding: 20px;
   display: flex;
   flex-direction: column;
   align-items: center;
-  gap: 30px;
+  gap: 20px;
+  @media (max-width: 768px) {
+    gap: 10px;
+  }
+  @media (min-width: 768px) and (max-width: 1024px) {
+    gap: 20px;
+  }
 }
 
 .icon {
   width: 70px;
   height: 70px;
+}
+
+.cardTitle {
+  font-weight: 500;
+  text-transform: uppercase;
+  @media (max-width: 768px) {
+    margin-bottom: 5px;
+  }
+}
+
+.cardText {
+  margin-bottom: 0px;
+}
+
+.cardText::first-letter {
+  text-transform: capitalize;
 }
