--- src/css/Carousel.css
@@ -1,56 +1,0 @@
-.carousel__body {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-}
-.body {
-  display: flex;
-  align-items: center;
-  flex: 1;
-  width: 100%;
-  height: 60vh;
-  margin-top: 50px;
-  flex-wrap: wrap;
-}
-
-.body__carousel {
-  flex: 0.8;
-  height: 100%;
-}
-.body__sidebar {
-  display: flex;
-  flex-direction: column;
-  flex: 0.1;
-  align-items: flex-end;
-  height: 100%;
-  justify-content: flex-start;
-}
-.paper__footerIcon {
-  width: 40px !important;
-  height: 40px !important;
-  margin-top: 5px;
-  color: black;
-  margin-right: 25px;
-}
-.sidebar_text {
-  transform: rotate(-90deg);
-  font-family: Oblik;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 24px;
-  word-break: keep-all;
-  margin-top: 50px;
-  overflow: hidden;
-}
-@media only screen and (max-width: 1270px) {
-  .sidebar_text {
-    transform: rotate(-90deg);
-    font-family: Oblik;
-    font-style: normal;
-    font-weight: bold;
-    font-size: 24px;
-    word-break: keep-all;
-    margin-top: -20px;
-    overflow: hidden;
-  }
-}
