+++ src/views/homePage/BookingWrapper.css
@@ -50,38 +50,4 @@
     width: 220px;
     overflow: hidden;
   }
-  .buttons button,
-  .buttons a {
-    color: black;
-    text-transform: uppercase;
-    width: 100%;
-    height: 50px;
-    border: 1px solid #04376f;
-    background: none;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-  .buttons .inline {
-    background: #04376f;
-    color: white;
-    outline: none;
-  }
-  .buttons .inline:hover {
-    background: white;
-    color: #04376f;
-  }
-  .buttons .outline {
-    background: transparent;
-    color: #04376f;
-    outline: none;
-  }
-  .buttons .outline:hover {
-    background: #04376f;
-    color: white;
-  }
-  .buttons .inline:focus,
-  .buttons .outline:focus {
-    outline: none;
-  }
 }
